Roll foward of changing global step to int64.
PiperOrigin-RevId: 169296498

+++ tensorflow/contrib/tpu/python/tpu/tpu_estimator.py
@@ -72,7 +72,7 @@ def _create_global_step(graph):
     return variable_scope.get_variable(
         ops.GraphKeys.GLOBAL_STEP,
         shape=[],
-        dtype=dtypes.int32,
+        dtype=dtypes.int64,
         initializer=init_ops.zeros_initializer(),
         trainable=False,
         use_resource=True,
